backend/requirements.txt
# ones-ai-platform backend dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
asyncpg==0.29.0
asyncssh==2.14.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.6.1
pydantic-settings==2.1.0
httpx==0.27.0
cryptography==42.0.5
openpyxl==3.1.2
